+++ src/pages/api/Videos/[videoId].tsx
@@ -1,5 +1,6 @@
 import dbConnect from '@/utils/dbConnect';
 import Videos from '@/models/Videos';
+import { getToken } from 'next-auth/jwt';
 dbConnect();
 export default async (req:any,res:any)=>{
     const videoId=req.query.videoId;
@@ -8,6 +9,10 @@ export default async (req:any,res:any)=>{
 
         case 'DELETE':
             try{
+                const session:any=await getToken({req});
+                if(!session || session.user.role!=="admin"){
+                    return res.status(401).json({message:"unauthorized"})
+                }
                 await Videos.findByIdAndDelete(videoId);
                 return res.status(200).send('delted');
             }
